user-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart